+++ internal/venice-test-common/src/main/java/com/linkedin/venice/utils/DataProviderUtils.java
@@ -164,19 +164,6 @@ public class DataProviderUtils {
     return allPermutationGenerator(IngestionTaskReusableObjects.Strategy.values());
   }
 
-  @DataProvider(name = "skipAdminOptions")
-  public Object[][] skipAdminOptions() {
-    return new Object[][] {
-        // 1) execution-id only -> should pass
-        { new String[] { "--execution-id", "10" }, false },
-        // 2) offset only -> should pass
-        { new String[] { "--offset", "10" }, false },
-        // 3) neither provided -> should fail
-        { new String[] {}, true },
-        // 4) both provided -> should fail
-        { new String[] { "--offset", "10", "--execution-id", "10" }, true } };
-  }
-
   /**
    * Generate permutations to be fed to a DataProvider.
    * For two boolean's we'd pass in allPermutationGenerator(BOOLEAN, BOOLEAN)
